requirements.txt
pyrogram>=2.0.0
TgCrypto
python-dotenv
sqlalchemy